refactor: divide to function components
